Move `colour_clf_io.ExponentStyle`, `colour_clf_io.LogStyle` and `colour_clf_io.RangeStyle` enumerations to `colour_clf_io.values` module.
